pyproject.toml
[tool.poetry]
name = "gpt-assistant"
version = "0.1.0"
description = "Atom the GPT Assistant is a smart home assistant powered with AI"
authors = ["Philip Yu <ht.yu@me.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gpt_assistant"}]

[tool.poetry.dependencies]
python = "3.9.15"
numpy = "1.23.5"
numba = "0.56.4"
sounddevice = "0.4.6"
pydub = "0.25.1"
scipy = "1.10.1"
boto3 = "1.26.115"
openai = "0.27.4"
python-dotenv = "1.0.0"
falcon = "3.1.1"
gunicorn = "20.1.0"
langchain = "0.0.149"
pvporcupine = "2.2.0"
pvrecorder = "1.1.1"
SpeechRecognition = "3.10.0"
poetry-setup = "^0.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
